requirements.txt
cadquery
openai
python-dotenv
streamlit
fastapi
uvicorn
